+++ src/main/java/roomescape/time/repository/JdbcReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.time;
+package roomescape.time.repository;
 
 import java.sql.PreparedStatement;
 import java.time.LocalTime;
@@ -9,9 +9,10 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.time.ReservationTime;
 
 @Repository
-public class ReservationTimeRepository {
+public class JdbcReservationTimeRepository implements ReservationTimeRepository {
 
     private final JdbcTemplate jdbcTemplate;
     private final RowMapper<ReservationTime> reservationTimeRowMapper = (resultSet, rowNum) -> new ReservationTime(
@@ -19,10 +20,11 @@ public class ReservationTimeRepository {
             resultSet.getObject("start_at", LocalTime.class)
     );
 
-    public ReservationTimeRepository(JdbcTemplate jdbcTemplate) {
+    public JdbcReservationTimeRepository(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
+    @Override
     public ReservationTime save(ReservationTime reservationTime) {
         String sql = "INSERT INTO reservation_time(start_at) VALUES(?)";
         KeyHolder keyHolder = new GeneratedKeyHolder();
@@ -36,17 +38,20 @@ public class ReservationTimeRepository {
         return new ReservationTime(id, reservationTime.getStartAt());
     }
 
+    @Override
     public List<ReservationTime> findAll() {
         String sql = "SELECT * FROM reservation_time";
         return jdbcTemplate.query(sql, reservationTimeRowMapper);
     }
 
+    @Override
     public Optional<ReservationTime> findById(Long id) {
         String sql = "SELECT * FROM reservation_time WHERE id = ?";
         List<ReservationTime> reservationTimes = jdbcTemplate.query(sql, reservationTimeRowMapper, id);
         return reservationTimes.stream().findFirst();
     }
 
+    @Override
     public void deleteById(Long id) {
         String sql = "DELETE FROM reservation_time WHERE id = ?";
         jdbcTemplate.update(sql, id);
